Consistently use uint_fast32_t for dag_len
rsort operates on uint_fast32_t, but everywhere else uses size_t.
This leasts to trucating casts on systemts where uint_fast32_t is uint32_t
and size_t is uint64_t.

Here we go through to try to make everything related to dag_len and dag indexing
consistently uint_fast32_t.

This is difficult to test.  I've tested it by making my own include/mystdint.h
which contains:

    #include <stdint.h>

    #define uint_fast32_t uint32_t
    #define int_fast32_t int32_t
    #define uint_fast16_t uint16_t
    #define int_fast16_t int16_t
    #undef PRIdFAST32
    #define PRIdFAST32 "d"

and replacing all other `#include <stdint.h>` with `#include "mystdint.h"`.

+++ C/dag.h
@@ -115,7 +115,7 @@ typedef struct dag_node {
   jet_ptr jet;
   sha256_midstate cmr;
   union {
-    size_t aux; /* Used as scratch space for verifyCanonicalOrder. */
+    uint_fast32_t aux; /* Used as scratch space for verifyCanonicalOrder. */
     struct {
        size_t sourceType, targetType;
     };
@@ -125,7 +125,7 @@ typedef struct dag_node {
       size_t sourceIx;
     };
     struct {
-      size_t child[2];
+      uint_fast32_t child[2];
     };
     bitstring compactValue;
   };
@@ -334,7 +334,7 @@ typedef struct analyses {
  * Precondition: dag_node dag[i + 1] and 'dag' is well-formed.
  *               dag[i].'tag' \notin {HIDDEN, JET, WORD}
  */
-void simplicity_computeCommitmentMerkleRoot(dag_node* dag, size_t i);
+void simplicity_computeCommitmentMerkleRoot(dag_node* dag, uint_fast32_t i);
 
 /* Given a well-typed dag representing a Simplicity expression, compute the annotated Merkle roots of all subexpressions.
  * For all 'i', 0 <= 'i' < 'len', 'analysis[i].annotatedMerkleRoot' will be the AMR of the subexpression denoted by the slice
@@ -347,7 +347,7 @@ void simplicity_computeCommitmentMerkleRoot(dag_node* dag, size_t i);
  *               dag_node dag[len] and 'dag' has witness data and is well-typed with 'type_dag'.
  * Postconditon: analyses analysis[len] contains the annotated Merkle roots of each subexpressions of 'dag'.
  */
-void simplicity_computeAnnotatedMerkleRoot(analyses* analysis, const dag_node* dag, const type* type_dag, size_t len);
+void simplicity_computeAnnotatedMerkleRoot(analyses* analysis, const dag_node* dag, const type* type_dag, uint_fast32_t len);
 
 /* Verifies that the 'dag' is in canonical order, meaning that nodes under the left branches have lower indices than nodes under
  * right branches, with the exception that nodes under right braches may (cross-)reference identical nodes that already occur under
@@ -360,7 +360,7 @@ void simplicity_computeAnnotatedMerkleRoot(analyses* analysis, const dag_node* d
  *
  * Precondition: dag_node dag[len] and 'dag' is well-formed.
  */
-simplicity_err simplicity_verifyCanonicalOrder(dag_node* dag, const size_t len);
+simplicity_err simplicity_verifyCanonicalOrder(dag_node* dag, const uint_fast32_t len);
 
 /* This function fills in the 'WITNESS' nodes of a 'dag' with the data from 'witness'.
  * For each 'WITNESS' : A |- B expression in 'dag', the bits from the 'witness' bitstring are decoded in turn
@@ -375,7 +375,7 @@ simplicity_err simplicity_verifyCanonicalOrder(dag_node* dag, const size_t len);
  * Postcondition: dag_node dag[len] and 'dag' has witness data and is well-typed with 'type_dag'
  *                  when the result is 'SIMPLICITY_NO_ERROR';
  */
-simplicity_err simplicity_fillWitnessData(dag_node* dag, type* type_dag, const size_t len, bitstream *witness);
+simplicity_err simplicity_fillWitnessData(dag_node* dag, type* type_dag, const uint_fast32_t len, bitstream *witness);
 
 /* Verifies that identity Merkle roots of every subexpression in a well-typed 'dag' with witnesses are all unique,
  * including that each hidden root hash for every 'HIDDEN' node is unique.
@@ -388,6 +388,6 @@ simplicity_err simplicity_fillWitnessData(dag_node* dag, type* type_dag, const s
  *
  * Precondition: dag_node dag[len] and 'dag' is well-typed with 'type_dag' and contains witnesses.
  */
-simplicity_err simplicity_verifyNoDuplicateIdentityRoots(sha256_midstate* imr, const dag_node* dag, const type* type_dag, const size_t dag_len);
+simplicity_err simplicity_verifyNoDuplicateIdentityRoots(sha256_midstate* imr, const dag_node* dag, const type* type_dag, const uint_fast32_t dag_len);
 
 #endif
