+++ C/rsort.c
@@ -267,7 +267,7 @@ int simplicity_hasDuplicates(const sha256_midstate* a, uint_fast32_t len) {
   static_assert(sizeof(a->s) * CHAR_BIT == 256, "sha256_midstate.s has unnamed padding.");
   static_assert(DAG_LEN_MAX <= UINT32_MAX, "DAG_LEN_MAX does not fit in uint32_t.");
   static_assert(DAG_LEN_MAX <= SIZE_MAX / sizeof(const sha256_midstate*), "perm array too large.");
-  simplicity_assert(len <= SIZE_MAX / sizeof(const sha256_midstate*));
+  simplicity_assert((size_t)len <= SIZE_MAX / sizeof(const sha256_midstate*));
   const sha256_midstate **perm = simplicity_malloc(len * sizeof(const sha256_midstate*));
   uint32_t *stack = simplicity_malloc(((CHAR_COUNT - 1)*(sizeof((*perm)->s)) + 1) * sizeof(uint32_t));
   int result = perm && stack ? 0 : -1;
